+++ test/com/intellij/advancedExpressionFolding/FoldingTemporaryTestEditor.kt
@@ -3,6 +3,7 @@ package com.intellij.advancedExpressionFolding
 import com.intellij.advancedExpressionFolding.diff.FoldingDescriptorExWrapper
 import com.intellij.advancedExpressionFolding.diff.FoldingTemporaryEditor
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.util.Computable
 import com.intellij.testFramework.runInEdtAndGet
 
 object FoldingTemporaryTestEditor {
@@ -10,13 +11,10 @@ object FoldingTemporaryTestEditor {
         text: String,
         wrapper: FoldingDescriptorExWrapper
     ): String {
-        var changedText: String? = null
-        //TODO: cant pass as reference, because its inline function
-        runInEdtAndGet {
-            ApplicationManager.getApplication().runWriteAction {
-                changedText = FoldingTemporaryEditor.foldInEditor(text, wrapper.list)
-            }
+        return runInEdtAndGet {
+            ApplicationManager.getApplication().runWriteAction(Computable {
+                FoldingTemporaryEditor.foldInEditor(text, wrapper.list)
+            })
         }
-        return changedText!!
     }
 }