+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 - [logFolding] - Enable LogBrackets folding for every method and support String.format with %s placeholders
 - [patternMatchingInstanceof] - Pattern Matching for instanceof (JEP 394)
 
+## [3.0.0] - 2025-03-03
+
+### Added
+
+- [logFolding] - Enable LogBrackets folding for every method and support String.format with %s placeholders
+- [patternMatchingInstanceof] - Pattern Matching for instanceof (JEP 394)
+
 ## [2.0.0] - 2024-09-15
 
 ### Added